Cache BG layer data per-scanline
The scanline cache is invalidated on X=0 and rebuilt lazily: When
lookup_bg_color is called and the scanline cache for the BG layer isn't
valid, we render the rest of the current scanline into the cache.

This had a little bit disappointing result, in that #9 still applies
(ie. it's not as fast as I hoped). However, the benchmark dropped from
about 10 ms per frame to a little bit more than 6 ms per frame, which is
still something.